rename syncService package to backgroundSync
